more code in seek.hpp
